backend/requirements.txt
flask
flask-cors
python-dotenv
flask-mysql
flask-mysqldb
flask-sqlalchemy
